created databse module to help with working with localStorage

+++ public/scripts/_helpers.mjs
@@ -8,33 +8,10 @@ export function removeAllChildNodes(parent) {
     }
 }
 
-export function arrayRemove(arr, value) {
-  const index = arr.indexOf(value);
-  arr.splice(index, 1);
-  return arr;
-}
-
 export function restartJavascriptDeliverable() {
-  localStorage.clear();
-  window.location.replace('index.html');
-}
-
-export async function getManyToOneDataFromDatabase(manyName, oneID) {
-  const manyIDsString = localStorage.getItem(`${oneID}_${manyName}s`) ?? "[]";
-  const manyIDs = JSON.parse(manyIDsString);
-  return manyIDs.map(id => {
-    const data = JSON.parse(localStorage.getItem(id))
-    data.id = id;
-    return data;
-  })
-}
-
-export async function getManyFromDatabase(manyName) {
-  const manyIDsString = localStorage.getItem(`${manyName}s`) ?? "[]";
-  const manyIDs = JSON.parse(manyIDsString);
-  return manyIDs.map(id => {
-    const data = JSON.parse(localStorage.getItem(id))
-    data.id = id;
-    return data;
-  })
+  const confirmation = confirm('Are you sure you want to restart the entire Javascript Deliverable? This will delete the entire testing database.')
+  if (confirmation) {
+    localStorage.clear();
+    window.location.replace('index.html');
+  }
 }
